Honor the cookie expiry setting for oauth
Currently if one sets the cookie_expiry value to a number then the
`session_expiry` will be used to set the `cookie expiry`. This can be
problem since if session_expiry is not a number then cookies will be set
to expire immediately. Therefore, if one wants to set the cookie expiry
today they must set both the `cookie_expiry` and `session_expiry`
together.