Reuse refesh_token if not provided an updated one
The spec seems to be a bit unclear here since the `refresh_token` is
optional in the repsonse. It seems to be open to interpretation whether
that means the original `refresh_token` is still valid. I think most
providers do allow reuse of `refresh_tokens` up until the expiry of that
token.

Should they not be reusable then that should just mean we fail to
refresh next time around and assertions should fail causing a 302.

See #44