Unpack knex response correctly when using pg client
Under https://knexjs.org/#Raw we are told `Note that the response will be whatever the underlying sql library would typically return on a normal query, so you may need to look at the documentation for the base library the queries are executing against to determine how to handle the response.`.

My assumption is that the existing response handler in the eas SQL adapter only supports the mysql database library considering the documentation is oriented towards mysql.

My proposed change allows us to correctly unpack a pg database library response while keeping the existing approach for all other libraries. The above change may need to be extended for the sqlite3 and mssql libraires depending on their specific response objects.